gantry-main/requirements.txt
Flask
flask-cors
pyserial
numpy
opencv-python
python-aravis